[ADD] Inventory usable for testing all purposes
This file contains the essentials of the ips as well as the groups associated with them in order to perform the necessary tests on the Ansible scripts. It also contains the local IPs of the virtual machines. To be used with care to avoid damaging the machines.
This one is in textual format. it can be used directly or reformatted into a .yaml or .json format.